requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
pandas
plotly
sentence-transformers